refactor Inversify stuff to be more modular
