+++ src/ioc/index.ts
@@ -1,37 +1,20 @@
 import { Server } from 'https';
-import { Container } from 'inversify';
 import { Browser } from 'puppeteer';
 
-import { BrowserService } from './browser/browser.service';
-import providePuppeteer from './browser/puppeteer.provider';
-import provideExpress from './host/express.provider';
-import { HostService } from './host/host.service';
-import { Logger } from './logger/logger.service';
-import { Renderer } from './renderer/renderer';
-import { IOptions, TYPES } from './types';
+import LifecycleContainer from './lifecycle-container';
 
-type TsClass = new (...args: any[]) => any;
-type InjectIdentifier = TsClass | symbol;
+import { BrowserService } from '../browser/browser.service';
+import providePuppeteer from '../browser/puppeteer.provider';
+import provideExpress from '../host/express.provider';
+import { HostService } from '../host/host.service';
+import { Logger } from '../logger/logger.service';
+import { Renderer } from '../renderer/renderer';
+import { IOptions, Provider, TYPES } from '../types';
 
-type Func = (...args: any[]) => any;
-
-export type Provider<T> = (...args: any[]) => Promise<T>;
-
-class CloseableContainer extends Container {
-  async ready(): Promise<void> {}
-  close() {}
-
-  bindDependencies(func: Func, dependencies: InjectIdentifier[]) {
-    const injections = dependencies.map((dependency: InjectIdentifier) => this.get(dependency));
-    return func.bind(func, ...injections);
-  }
-}
-
-function context(options: IOptions) {
-  // Create extended container with lifecycle management hooks
-  const container = new CloseableContainer();
-
-  // Bind classes & values
+/*
+ * Bind classes & values to our container
+ */
+function bindings(container: LifecycleContainer, options: IOptions) {
   container.bind<IOptions>(TYPES.Options).toConstantValue(options);
   container
     .bind<HostService>(HostService)
@@ -49,14 +32,30 @@ function context(options: IOptions) {
     .bind<Logger>(Logger)
     .toSelf()
     .inSingletonScope();
+}
 
-  // Bind providers
+/*
+ * Bind providers to container
+ */
+function providers(container: LifecycleContainer) {
   const serverProvider = container.bindDependencies(provideExpress, [TYPES.Options, Logger]);
   const browserProvider = container.bindDependencies(providePuppeteer, [Logger]);
 
   container.bind<Provider<Server>>(TYPES.ServerProvider).toProvider<Server>(() => serverProvider);
 
   container.bind<Provider<Browser>>(TYPES.BrowserProvider).toProvider<Browser>(() => browserProvider);
+}
+
+/*
+ * Create an application context, passing it the CLI options we get
+ */
+function context(options: IOptions) {
+  // Create extended container with lifecycle management hooks
+  const container = new LifecycleContainer();
+
+  // Do bindings
+  bindings(container, options);
+  providers(container);
 
   // Implement container lifecycle hooks
   container.ready = async function() {
